service-b2b-data/pyproject.toml
[tool.poetry]
name = "service-b2b-data"
version = "0.1.0"
description = "Microservice for company data verification with Redis cache"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
websockets = "^12.0"
redis = "^5.0.1"
httpx = "^0.26.0"
fastapi-mail = "^1.4.1"
pydantic-settings = "^2.1.0"
langchain = "^0.1.0"
langchain-google-genai = "^0.0.9"
pgvector = "^0.2.4"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.2.1"
mypy = "^1.8.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["B008"]


[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
